WIP: refactor header/sidebar
