+++ frontend/src/components/ui/layout/DashboardLayoutRoute.tsx
@@ -1,0 +1,11 @@
+// src/components/layout/DashboardLayoutRoute.tsx
+import { Outlet } from 'react-router-dom';
+import DashboardLayout from './DashboardLayout';
+
+export default function DashboardLayoutRoute() {
+  return (
+    <DashboardLayout>
+      <Outlet />
+    </DashboardLayout>
+  );
+}
